各種Socketクラスの構成を変更した。
