+++ src/main/java/com/daichi703n/echo/EchoApplication.java
@@ -2,11 +2,9 @@ package com.daichi703n.echo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableAsync;
 
 
 @SpringBootApplication
-@EnableAsync
 public class EchoApplication {
 
     public static void main(String[] args) {
